# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Upload Spreadsheet API V1 Documentation",
        "description": "Upload Spreadsheet API V1 Documentation",
        "contact": {
            "email": "ayodeleoniosun63@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Test Server"
        }
    ],
    "paths": {
        "/contracts/upload": {
            "post": {
                "tags": [
                    "Contracts"
                ],
                "summary": "Upload contract",
                "description": "Upload contract",
                "operationId": "763f8984cc8146b17b4e1b6a6f2d921a",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "file to upload",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "The response code",
                                            "type": "string"
                                        },
                                        "message": {
                                            "description": "The response message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "The response code",
                                            "type": "string"
                                        },
                                        "message": {
                                            "description": "The response message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "An error occured"
                    }
                }
            }
        },
        "/contracts": {
            "get": {
                "tags": [
                    "Contracts"
                ],
                "summary": "Get list of contracts",
                "description": "Returns list of contracts",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "The response code",
                                            "type": "string"
                                        },
                                        "contracts": {
                                            "description": "The response data",
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "500": {
                        "description": "An error occured."
                    }
                }
            }
        },
        "/contracts?search={searchTerm}": {
            "get": {
                "tags": [
                    "Contracts"
                ],
                "summary": "Search contract by date (dataCelebracaoContrato), amount (range) (precoContratual), winning company (adjudicatarios)",
                "description": "Returns search result",
                "operationId": "58e1473ba827a640dfed373e1b267169",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "search term",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "The response code",
                                            "type": "string"
                                        },
                                        "contracts": {
                                            "description": "The response data",
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "500": {
                        "description": "An error occured."
                    }
                }
            }
        },
        "/contracts/{id}": {
            "get": {
                "tags": [
                    "Contracts"
                ],
                "summary": "Find contract",
                "description": "Returns details of a single contract",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "contract id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "The response code",
                                            "type": "string"
                                        },
                                        "contract": {
                                            "description": "The response data",
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "500": {
                        "description": "An error occured."
                    }
                }
            }
        },
        "/contracts/{id}/status": {
            "get": {
                "tags": [
                    "Contracts"
                ],
                "summary": "Find contract status",
                "description": "Returns the status of a single contract",
                "operationId": "1b32cced63fdb88129166c21153128df",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "contract id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "The response code",
                                            "type": "string"
                                        },
                                        "contract": {
                                            "description": "The response data",
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "500": {
                        "description": "An error occured."
                    }
                }
            }
        }
    }
}